Create pythonvenv
This is a script I created which was inspired by a series of chats with senior programmers I met while using Libera, a well-established IRC support community. Based on the advice shared with me, I created a Bash script which can be used to quickly generate a Python venv which works great if you want to work on Python code using a Linux system such as Ubuntu or Red Hat. Good luck to all of those who read this and find it helpful. There are other tools to get the job done but we all prefer our own. I hope my own tools such as this one help you, whomever you are and wherever you sit, get the job done.